requirements.txt
python-telegram-bot==13.15
httpx==0.21.0